backend1/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pillow
pytesseract
requests
python-dotenv
